==> analysis/2-merge-with-census.ipynb <==
# Join on Census data

Using tidycensus, this notebook grabs zip-code level demographic information and adds to the aggregated school-level data.

## Setup Python and R environment
you can ignore this section

%load_ext rpy2.ipython
%load_ext autoreload
%autoreload 2

%matplotlib inline  
from matplotlib import rcParams
rcParams['figure.figsize'] = (16, 100)

import warnings
from rpy2.rinterface import RRuntimeWarning
warnings.filterwarnings("ignore") # Ignore all warnings
# warnings.filterwarnings("ignore", category=RRuntimeWarning) # Show some warnings

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from IPython.display import display, HTML

%%javascript
// Disable auto-scrolling
IPython.OutputArea.prototype._should_scroll = function(lines) {
    return false;
}

%%R

# My commonly used R imports

require('tidyverse')
require(data.table)


## Load & Clean Data

👉 Load the data along with the census connectors below (the output of the `connect-to-census.ipynb` notebook) and do any cleanup you'd like to do.

%%R

df <- fread("../data/2021-22-crdc-data/merged_data.csv")

%%R

df %>% head(5)

## 👉 Grab Census Data

1. loading the Census API key

import dotenv

# Load the environment variables
# (loads CENSUS_API_KEY from .env)
dotenv.load_dotenv()


%%R 

require('tidycensus')

# because it an environment variable, we don't have to 
# explicitly pass this string to R, it is readable here
# in this R cell.
census_api_key(Sys.getenv("CENSUS_API_KEY"))

2. Decide which Census variables you want

    Use <https://censusreporter.org/> to figure out which tables you want. (if censusreporter is down, check out the code in the cell below)

    -   Scroll to the bottom of the page to see the tables.
    -   If you already know the table ID, stick that in the "Explore" section to learn more about that table.

    By default this code loads (B01003_001) which we found in censusreporter here: https://censusreporter.org/tables/B01003/

    - find some other variables that you're also interested in
    - don't forget to pick a geography like "tract", "county" or "block group". here is the list of [all geographies](https://walker-data.com/tidycensus/articles/basic-usage.html#geography-in-tidycensus
    ).


%%R 

# Finding the Census Varaibles for the ACS 5 year survey
# Generall you'd do this in CensusReporter, but since it's down sometimes, here it is using tidycensus's load_variables function

# get every single variable in the ACS5
all_census_vars <- load_variables(2021, "acs5", cache = TRUE) 

filtered_census_vars <- all_census_vars %>% 
    filter(grepl("income", label, ignore.case = TRUE))   # filter to those containing "median income"
    
# write to CSV so we can view it in python
filtered_census_vars %>% 
    write_csv("../data/filtered_census_vars.csv")

# show the first few rows
filtered_census_vars %>%
    select(-geography) %>% # remove the geography column
    print(n = 20) # print the first 20 rows

%%R 
# the variable B01003_001E was selectd from the census table 
# for population, which we found in censusreporter here:
# https://censusreporter.org/tables/B01003/

# in the table below, pick the geography, the variables, and the survey you want to pull from
# see the possible values here https://walker-data.com/tidycensus/articles/basic-usage.html

# Get variable from ACS
census_data <- get_acs(geography = "zcta", 
                      variables = c(
                        population="B01003_001",
                        med_hh_inc="B19013_001",
                        med_fam_inc="B19113_001",
                        med_fam_inc_w_children="B19125_002",
                        gini_inequality="B19083_001",
                        poverty_pop="B17001_002",
                        white_pop ="B02001_002",
                        total_racial_pop="B02001_001",
                        total_edu_pop="B15003_001",
                        bach_pop="B15003_022",
                        masters_pop="B15003_023",
                        phd_pop="B15003_025",
                        prof_school_pop="B15003_024"
                      ), 
                      year = 2021,
                      survey="acs5",
                      geometry=T)

census_data

%%R 

pivoted_census <- census_data %>%
    select(-moe) %>%
    pivot_wider(names_from = "variable", values_from="estimate") %>%
    mutate(poverty_prop = poverty_pop / population,
           white_prop = white_pop / total_racial_pop,
           bach_plus_prop = (bach_pop + masters_pop + phd_pop + prof_school_pop) / total_edu_pop)



%%R
library(stringr)

cleaned_census <- pivoted_census %>% mutate(zip_code = as.double(str_sub(NAME, -5,-1)))

## 👉 Merge it with your data

hint...`tidycensus` provides you data in long format you may need to pivot the census data from long to wide format before merging it with your data

%%R


school_census_data <- df %>% left_join(cleaned_census, by=join_by(LEA_ZIP == zip_code))

%%R 

write_csv(school_census_data, "../data/final_data.csv")
